+++ src/main/java/br/com/zalf/prolog/webservice/frota/pneu/afericao/configuracao/_model/ConfiguracaoTipoVeiculoAferivelListagem.java
@@ -2,6 +2,7 @@ package br.com.zalf.prolog.webservice.frota.pneu.afericao.configuracao._model;
 
 import lombok.Data;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Created on 03/05/18.
@@ -11,23 +12,28 @@ import org.jetbrains.annotations.NotNull;
 @Data
 public final class ConfiguracaoTipoVeiculoAferivelListagem {
 
-    @NotNull
+    @Nullable
     private final Long codConfiguracao;
     @NotNull
     private final Long codUnidade;
     @NotNull
     private final ConfiguracaoTipoVeiculoAferivelVeiculoVisualizacao tipoVeiculo;
-    private final boolean podeAferirPressao;
-    private final boolean podeAferirSulco;
-    private final boolean podeAferirSulcoPressao;
     private final boolean podeAferirEstepe;
     @NotNull
     private final FormaColetaDadosAfericaoEnum formaColetaDadosPressao;
     @NotNull
+    private final String formaColetaDadosPressaoLegivel;
+    @NotNull
     private final FormaColetaDadosAfericaoEnum formaColetaDadosSulco;
+    @NotNull
+    private final String formaColetaDadosSulcoLegivel;
     @NotNull
     private final FormaColetaDadosAfericaoEnum formaColetaDadosSulcoPressao;
     @NotNull
+    private final String formaColetaDadosSulcoPressaoLegivel;
+    @NotNull
     private final FormaColetaDadosAfericaoEnum formaColetaDadosFechamentoServico;
+    @NotNull
+    private final String formaColetaDadosFechamentoServicoLegivel;
 
 }