+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Change Log
 ### Features
 * Permite parametrização de aferições e fechamento de serviço manuais, com equipamentos ou ambos (PL-2689)
 * Permite salvar forma de coleta de dados da aferição (PL-2686)
+* Permite salvar forma de coleta de dados no fechamento de serviços (PL-2714)
+* Busca a forma de coleta dos dados para os serviços fechados (PL-2715)
+
+## Refactors
+* Refatora objetos que utilizam booleans podeAferirSulco, Pressao e SulcoPressao para não utilizar ou serem adaptados (PL-2689)
+
+<a name="v1.0.30"></a>
+## Version [v1.0.30](https://github.com/luizfp/PrologWebService/compare/v1.0.29...v1.0.30) (release-date) [unreleased]
 
 <a name="v1.0.29"></a>
 ## Version [v1.0.29](https://github.com/luizfp/PrologWebService/compare/v1.0.28...v1.0.29) (2020-04-28)